+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Add support for Real-Time Collaboration (`jupyter-collaboration`).
+  (Set `collaboration.enable = true`.)
+
 ### Changed
 
 * Rename `jupyterExtensions` -> `labextensions`.